repo/hello_agent/agent run and global level/explore-uv/pyproject.toml
[project]
name = "explore-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ayman Rehman", email = "aimanshahzad295@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
]

[project.scripts]
explore-uv = "explore_uv:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = [
    "hello_agent",
    "hello_agent/hello_agent",
]
